pyproject.toml
[tool.poetry]
name = "wtcombot"
version = "0.0.1"
description = ""
authors = [
            "Arina Pankova <arinpank(at)gmail.com>",
            "Dmitry Danilov <dima.danilov(at)gmail.com>"
          ]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
requests = "^2.28.1"
requests-toolbelt = "^0.9.1"
heyoo = "^0.0.6"
pyTelegramBotAPI = "^4.7.1"
python-dotenv = "^0.21.0"
gunicorn = "^20.1.0"

#[tool.poetry.group.dev.dependencies]
#black = "^22.10.0"
#isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#[tool.isort]
#profile = "black"
#multi_line_output = 3
#include_trailing_comma = true
